requirements.txt
flywheel-sdk
pandas
pangres
psycopg2-binary
sqlalchemy
yaspin
